twister: Split output by lines before it is matched with regex
Harness console allows to define what output is required from a
device under test. However, a "single" output can contain multiple
lines. The commit split each output into individual lines and each
line is verified with pattern matching. This resolves an issue where
multiple lines are treated as a single output. In such case twister
were going to a next output as soon as it founds the first match,
and then was not able to recognize that the rest of required lines
were already printed in the resolved output.

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>